Add `:legacy:ui:theme` module and move `ThemeManager` and `ThemeProvider`
