+++ legacy/common/src/main/java/com/fsck/k9/CommonApp.kt
@@ -5,11 +5,11 @@ import android.content.res.Configuration
 import android.content.res.Resources
 import app.k9mail.feature.widget.message.list.MessageListWidgetManager
 import app.k9mail.legacy.di.DI
+import app.k9mail.legacy.ui.theme.ThemeManager
 import com.fsck.k9.controller.MessagingController
 import com.fsck.k9.job.WorkManagerConfigurationProvider
 import com.fsck.k9.notification.NotificationChannelManager
 import com.fsck.k9.ui.base.AppLanguageManager
-import com.fsck.k9.ui.base.ThemeManager
 import com.fsck.k9.ui.base.extensions.currentLocale
 import java.util.Locale
 import kotlinx.coroutines.CoroutineScope
