+++ legacy/ui/base/src/main/java/com/fsck/k9/ui/base/KoinModule.kt
@@ -1,5 +1,6 @@
 package com.fsck.k9.ui.base
 
+import app.k9mail.legacy.ui.theme.ThemeManager
 import com.fsck.k9.ui.base.locale.SystemLocaleManager
 import org.koin.core.qualifier.named
 import org.koin.dsl.module
