+++ legacy/ui/legacy/src/main/java/com/fsck/k9/ui/K9ThemeProvider.kt
@@ -1,6 +1,6 @@
 package com.fsck.k9.ui
 
-import com.fsck.k9.ui.base.ThemeProvider
+import app.k9mail.legacy.ui.theme.ThemeProvider
 
 // TODO: Move this class and the theme resources to the main app module
 class K9ThemeProvider : ThemeProvider {
