+++ legacy/ui/legacy/src/main/java/com/fsck/k9/ui/helper/HtmlSettingsProvider.kt
@@ -1,9 +1,9 @@
 package com.fsck.k9.ui.helper
 
+import app.k9mail.legacy.ui.theme.Theme
+import app.k9mail.legacy.ui.theme.ThemeManager
 import com.fsck.k9.K9
 import com.fsck.k9.message.html.HtmlSettings
-import com.fsck.k9.ui.base.Theme
-import com.fsck.k9.ui.base.ThemeManager
 
 class HtmlSettingsProvider(private val themeManager: ThemeManager) {
     fun createForMessageView() = HtmlSettings(
